+++ packages/extended-angular/src/app/views/components/count-card/count-card-routing.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { CountCardDemoComponent } from './count-card.component';
+
+const routes: Routes = [{ path: '', component: CountCardDemoComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class CountCardRoutingModule {}